+++ app/javascript/packs/quizBankPage/NewQuizQuestionForm.jsx
@@ -20,12 +20,11 @@ const NewQuizQuestionForm = ({type, marks, setMarks}) => {
   function changeLocalMarks(e) {
     e.preventDefault()
     if (!isNaN(+e.target.value)) {
-      console.log("This is an integer")
+      // This is an integer
       let disparity = e.target.value - localMark
       setMarks(marks + disparity)
       setLocalMark(e.target.value)
     } else {
-      console.log("This is not an integer")
     }
   }